Fixes #151:
* Merged `device_count.hpp` into `miscellany.hpp`
