+++ src/cuda/api/miscellany.hpp
@@ -1,13 +1,15 @@
 /**
  * @file miscellany.hpp
  *
- * @brief Miscellaneous functionality which does not fit in another file
+ * @brief Miscellaneous functionality which does not fit in another file,
+ * and does not depend on the main proxy classes
  *
  */
 #ifndef CUDA_API_WRAPPERS_MISCELLANY_HPP_
 #define CUDA_API_WRAPPERS_MISCELLANY_HPP_
 
-#include "error.hpp"
+#include <cuda/api/types.hpp>
+#include <cuda/api/error.hpp>
 
 #include <cuda_runtime_api.h>
 
@@ -29,6 +31,39 @@ void force_runtime_initialization()
 	auto status = cudaFree(nullptr);
 	throw_if_error(status, "Forcing CUDA runtime initialization");
 }
+
+namespace device {
+
+/**
+ * Get the number of CUDA devices usable on the system (with the current CUDA
+ * library and kernel driver)
+ *
+ * @note This _should_ be returning an unsigned value; unfortunately, device::id_t  is
+ * signed in CUDA for some reason and we maintain compatibility (although this might
+ * change in the future). So... the returned type is the same as in cudaGetDeviceCount,
+ * a signed integer.
+ *
+ * @return the number of CUDA devices on this system
+ * @throws cuda::error if the device count could not be obtained
+ */
+inline device::id_t  count()
+{
+	int device_count = 0; // Initializing, just to be on the safe side
+	status_t result = cudaGetDeviceCount(&device_count);
+	if (result == status::no_device) {
+		return 0;
+	}
+	else {
+		throw_if_error(result, "Failed obtaining the number of CUDA devices on the system");
+	}
+	if (device_count < 0) {
+		throw std::logic_error("cudaGetDeviceCount() reports an invalid number of CUDA devices");
+	}
+	return device_count;
+}
+
+} // namespace device
+
 
 } // namespace cuda
